Restructure agent spring configuration package to match web for consistency
Clear up what is autoconfig and what is environment processors. Make it explicit the whole package is spring specific.
