+++ genie-agent/src/main/java/com/netflix/genie/agent/spring/autoconfigure/ProcessAutoConfiguration.java
@@ -15,7 +15,7 @@
  *     limitations under the License.
  *
  */
-package com.netflix.genie.agent.configs;
+package com.netflix.genie.agent.spring.autoconfigure;
 
 import com.netflix.genie.agent.execution.process.JobProcessManager;
 import com.netflix.genie.agent.execution.process.impl.JobProcessManagerImpl;
