+++ genie-agent/src/main/java/com/netflix/genie/agent/spring/autoconfigure/package-info.java
@@ -17,12 +17,12 @@
  */
 
 /**
- * Any environment post processors for Spring that are needed.
+ * Spring Boot auto configuration classes.
  *
  * @author tgianos
  * @since 4.0.0
  */
 @ParametersAreNonnullByDefault
-package com.netflix.genie.agent.configs.processors;
+package com.netflix.genie.agent.spring.autoconfigure;
 
 import javax.annotation.ParametersAreNonnullByDefault;
